// src/api/model/serviceCard.ts
/**
 * Generated by orval v7.7.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { MaterialGroupRef } from './materialGroupRef';
import type { UnitEnum } from './unitEnum';
import type { ServiceTypeEnum } from './serviceTypeEnum';
import type { ServiceCardUnit } from './serviceCardUnit';

export interface ServiceCard {
  /** @nullable */
  id?: number | null;
  serviceCode: string;
  serviceName: string;
  serviceGroup: MaterialGroupRef;
  defaultUnit: UnitEnum;
  serviceType: ServiceTypeEnum;
  /** @nullable */
  optimalLevel?: number | null;
  /** @nullable */
  minimumLevel?: number | null;
  /** @nullable */
  specialCode?: string | null;
  /** @nullable */
  shelfLocation?: string | null;
  /** @nullable */
  serviceCardUnits?: ServiceCardUnit[] | null;
}